aoi.py: - use AOI class to manage download of data (rather than SingleBandItemEO object) - parse_input_raster returns a list no matter if single or multi-band rasters default_evaluate.yaml: add evaluate config metrics.py: add iou by torchmetrics

+++ evaluate_segmentation.py
@@ -3,13 +3,17 @@ import logging
 from pathlib import Path
 from typing import Sequence
 
+import omegaconf
+import torch
 from pandas.io.common import is_url
+from solaris import vector
+from solaris.utils.core import _check_rasterio_im_load
 from torch.hub import load_state_dict_from_url
 
 from dataset.aoi import aois_from_csv
-from utils.metrics import iou_per_obj
+from utils.metrics import iou_per_obj, iou_torchmetrics
 from utils.utils import get_key_def, gdl2pl_checkpoint, read_checkpoint, \
-    override_model_params_from_checkpoint
+    override_model_params_from_checkpoint, get_device_ids, set_device
 
 from inference_segmentation import main as gdl_inference
 from postprocess_segmentation import main as gdl_postprocess
@@ -35,6 +39,7 @@ def main(cfg):
     models_dir.mkdir(exist_ok=True)
     checkpoint = get_key_def('state_dict_path', cfg['inference'], expected_type=str, to_path=True, validate_path_exists=True)
     download_data = get_key_def('download_data', cfg['inference'], default=False, expected_type=bool)
+    min_iou = get_key_def('iou_threshold', cfg['evaluate'], default=0.5, expected_type=float)
             
     # Create yaml to use pytorch lightning model management
     if is_url(checkpoint):
@@ -45,7 +50,12 @@ def main(cfg):
     cfg_overridden = override_model_params_from_checkpoint(params=cfg.copy(), checkpoint_params=checkpoint_dict['params'])
 
     # Dataset params
-    bands_requested = get_key_def('bands', cfg_overridden['dataset'], default=("red", "blue", "green"), expected_type=Sequence)
+    bands_requested = get_key_def('bands', cfg_overridden['dataset'], default=("red", "green", "blue"), expected_type=Sequence)
+
+    # Set device for raster metrics
+    num_devices = get_key_def('num_gpus', cfg['training'], default=1)
+    gpu_devices_dict = get_device_ids(num_devices)
+    device = set_device(gpu_devices_dict=gpu_devices_dict)
 
     # read input data from csv
     benchmark_raw_data = aois_from_csv(
@@ -59,36 +69,78 @@ def main(cfg):
 
     metrics = []
     for aoi in benchmark_raw_data:
+        metric_per_aoi = {'state_dict': Path(checkpoint).name}
         logging.info(f"Benchmarking: {aoi.aoi_id}")
         # inference on each raster
         cfg['inference']['input_stac_item'] = aoi.raster_raw_input
         cfg['inference']['output_name'] = aoi.aoi_id + '_BUIL'
+
         pred_raster, pred_raster_path = gdl_inference(cfg.copy())
-        #print(pred_raster_path)
+        pred_raster = _check_rasterio_im_load(pred_raster_path).read()
+
+        # burn to raster
+        burn_val = 1 if not aoi.attr_field_filter and not aoi.attr_values_filter else None
+        label_raster = vector.mask.footprint_mask(
+            df=aoi.label_gdf,
+            reference_im=aoi.raster,
+            burn_field=aoi.attr_field_filter,
+            burn_value=burn_val)
+
         # compute raster metrics from prediction and ground truth
-        # TODO: where to reference numpy arrays?
-        #jaccard = JaccardIndex(num_classes=num_classes, ignore_index=None, threshold=0.5, multilabel=True)
+        # FIXME num_classes and ignore_index
+        iou = iou_torchmetrics(torch.from_numpy(pred_raster), torch.from_numpy(label_raster), device=device)
+        metric_per_aoi['iou_raster'] = iou
 
         # compute vector metrics from prediction and ground truth
         pred_vector_path = gdl_postprocess(cfg.copy())
+        #pred_vector_path = "tests/data/spacenet/SpaceNet_AOI_2_Las_Vegas-056155973080_01_P001-WV03_pred.gpkg"
 
-        metric = iou_per_obj(
+        metric_vector = iou_per_obj(
             pred=pred_vector_path,
             gt=aoi.label,
             attr_field=aoi.attr_field_filter,
             attr_vals=[cfg['dataset']['classes_dict']['BUIL']],
             aoi_id=aoi.aoi_id,
             aoi_categ=None,  # TODO: add category for human-readable report
+            min_iou=min_iou,
             gt_clip_bounds=None)
-        metric['state_dict'] = Path(checkpoint).name
-        metrics.append(metric)
-        break
+        metric_vector['iou_threshold'] = min_iou
+        metric_per_aoi.update(metric_vector)
+        metrics.append(metric_per_aoi)
     keys = metrics[0].keys()
     outpath = root / "benchmark.csv"
-    with open(outpath, 'w', newline='') as output_file:
-        dict_writer = csv.DictWriter(output_file, keys)
-        dict_writer.writeheader()
-        dict_writer.writerows(metrics)
+    logging.info(f"Benchmark report will be saved: ")
+    if not outpath.is_file():
+        with open(outpath, 'w', newline='') as output_file:
+            dict_writer = csv.DictWriter(output_file, keys)
+            dict_writer.writeheader()
+            dict_writer.writerows(metrics)
+    else:
+        with open(outpath, 'a', newline='') as output_file:
+            dict_writer = csv.DictWriter(output_file, keys)
+            dict_writer.writerows(metrics)
 
 
+if __name__ == '__main__':
+    # test benchmarking with spacenet data
+    pred_raster_path = "tests/data/spacenet/SpaceNet_AOI_2_Las_Vegas-056155973080_01_P001-WV03_pred.tif"
+    pred_raster = _check_rasterio_im_load(pred_raster_path).read()
 
+    label_raster_path = "tests/data/spacenet/SpaceNet_AOI_2_Las_Vegas-056155973080_01_P001-WV03_gt.tif"
+    label_raster = _check_rasterio_im_load(pred_raster_path).read()
+
+    iou = iou_torchmetrics(torch.from_numpy(pred_raster), torch.from_numpy(label_raster), device='cuda:0')
+    print(iou)
+
+    csv_file = "tests/sampling/sampling_segmentation_binary-stac_ci.csv"
+    checkpoint = "/media/data/ccmeo_models/pl_unet_resnet50_epoch-62-step-24066_gdl.ckpt"
+    cfg = {'dataset': {'raw_data_csv': csv_file,
+                       'classes_dict': {"BUIL": 1},
+                       'raw_data_dir': "/home/remi/PycharmProjects/geo-deep-learning/tests/data/spacenet"},
+           'inference': {'state_dict_path': checkpoint,
+                         'download_data': True},
+           'model': None,
+           'evaluate': {'iou_threshold': 0.5},
+           'training': {}}
+    cfg = omegaconf.DictConfig(cfg)
+    main(cfg)
